+++ priv/components/button.exs
@@ -33,8 +33,8 @@
       ],
       size: ["extra_small", "small", "medium", "large", "extra_large", "full"],
       rounded: ["extra_small", "small", "medium", "large", "extra_large", "full", "none"],
-      type: ["button_group", "button", "input_button", "button_link"],
-      only: ["button_group", "button", "input_button", "button_link"],
+      type: ["button_group", "button", "input_button", "button_link", "back"],
+      only: ["button_group", "button", "input_button", "button_link", "back"],
       helpers: [],
       module: ""
     ],