+++ priv/components/icon.exs
@@ -1,0 +1,13 @@
+[
+  icon: [
+    name: "icon",
+    args: [
+      type: ["icon"],
+      only: ["icon"],
+      helpers: [],
+      module: ""
+    ],
+    optional: [],
+    necessary: []
+  ]
+]