+++ priv/components/navbar.exs
@@ -21,7 +21,7 @@
       padding: ["extra_small", "small", "medium", "large", "extra_large", "none"],
       rounded: ["extra_small", "small", "medium", "large", "extra_large"],
       space: ["extra_small", "small", "medium", "large", "extra_large"],
-      only: ["navbar"],
+      only: ["navbar", "header"],
       helpers: [],
       module: ""
     ],