+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Features:
 - Add on_select attribute on tab slot
+- Add scroll area component [#296](https://github.com/mishka-group/mishka_chelekom/pull/296)
+- Add missing features of Phoenix core components [#296](https://github.com/mishka-group/mishka_chelekom/pull/296)
 
 ### Bugs:
 - Add missing slot to gallery component [#294](https://github.com/mishka-group/mishka_chelekom/pull/294)